requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
reportlab==4.4.3
Pillow==11.3.0
python-dotenv==1.1.1
